Move Attack to battle
