+++ battle/Battle.py
@@ -6,7 +6,7 @@ from random import choice as rand_choice
 from random import randint
 from typing import List, Tuple
 
-from Attack import Attack
+from battle.Attack import Attack
 from Cog import Cog
 from core.AttackGlobals import GROUP, MULTIPLIER, MULTIPLIER_DEFAULT
 from core.Entity import BattleEntity, Entity
